Add way to access species from ProvidedSpecies
This adds a way for the ProvidedSpecies enum to directly retrieve
the PokemonSpecies object it refers to. This ultimately reduces the
boilerplate for using species within the mod, as before you would need
to access the species repository for each provided species call just
to get the PokemonSpecies data.

While doing this, there were issues making the enum directly
inherit from the PokemonSpecies class (as it's an Enum)
and making PokemonSpecies an interface simply introduces
unnecessary complexity and boilerplate so it was decided against.

The outcome was creating a get method that fetches from the
repository.